This PR changes the `Equals` method of `NibblePath` to be a bit shorter than the `FindFirstDiff...` and delegate the majority to the `SequenceEqual`.

### Plan of Action

- [x] implement
- [x] benchmark various lengths
- [x] test various lengths

### Benchmarks

Do not show a lot beside the code size that is greatly reduced. Also, memory alignment may show up.

### Before

| Method                  | slice | length | Mean     | Error     | StdDev    | Code Size |
|------------------------ |------ |------- |---------:|----------:|----------:|----------:|
| Equals_not_equal_sliced | ?     | ?      | 3.664 ns | 0.0520 ns | 0.0487 ns |     643 B |
| Equals                  | 0     | 1      | 4.471 ns | 0.0713 ns | 0.0632 ns |     649 B |
| Equals                  | 0     | 16     | 4.489 ns | 0.0842 ns | 0.0788 ns |     649 B |
| Equals                  | 0     | 32     | 4.460 ns | 0.0590 ns | 0.0552 ns |     649 B |
| Equals                  | 0     | 48     | 4.447 ns | 0.0606 ns | 0.0537 ns |     649 B |
| Equals                  | 0     | 64     | 4.506 ns | 0.0986 ns | 0.1136 ns |     649 B |
| Equals                  | 1     | 1      | 4.420 ns | 0.0599 ns | 0.0531 ns |     649 B |
| Equals                  | 1     | 16     | 4.440 ns | 0.0946 ns | 0.0885 ns |     649 B |
| Equals                  | 1     | 32     | 4.498 ns | 0.0642 ns | 0.0601 ns |     649 B |
| Equals                  | 1     | 48     | 4.480 ns | 0.0589 ns | 0.0551 ns |     649 B |


### After 

| Method                  | slice | length | Mean     | Error     | StdDev    | Code Size |
|------------------------ |------ |------- |---------:|----------:|----------:|----------:|
| Equals_not_equal_sliced | ?     | ?      | 2.462 ns | 0.0492 ns | 0.0460 ns |     451 B |
| Equals                  | 0     | 1      | 4.433 ns | 0.0499 ns | 0.0467 ns |     452 B |
| Equals                  | 0     | 16     | 4.477 ns | 0.0622 ns | 0.0552 ns |     452 B |
| Equals                  | 0     | 32     | 4.473 ns | 0.0747 ns | 0.0699 ns |     452 B |
| Equals                  | 0     | 48     | 4.459 ns | 0.0667 ns | 0.0624 ns |     452 B |
| Equals                  | 0     | 64     | 4.530 ns | 0.0997 ns | 0.0933 ns |     452 B |
| Equals                  | 1     | 1      | 4.443 ns | 0.0446 ns | 0.0372 ns |     452 B |
| Equals                  | 1     | 16     | 4.481 ns | 0.1019 ns | 0.0953 ns |     452 B |
| Equals                  | 1     | 32     | 4.489 ns | 0.0611 ns | 0.0572 ns |     452 B |
| Equals                  | 1     | 48     | 4.434 ns | 0.0422 ns | 0.0374 ns |     452 B |