@Scooletz 2024-07-26
Merging as is. Does not show a lot in regards to the numbers, but it's much simpler and much smaller code, with a potential of getting inlined.